requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blessings==1.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
discord==2.3.2
discord.py==2.3.2
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.13.1
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
glfw==2.6.5
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
mdurl==0.1.2
mpmath==1.3.0
mujoco-py==2.1.2.14
multidict==6.0.5
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omgifol==0.5.1
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
progress==1.6
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.4.1
requests==2.31.0
rich==13.7.0
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tifffile==2023.7.10
torch==2.2.0
tqdm==4.66.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
yarl==1.9.4
zipp==3.17.0
